pyproject.toml
[tool.poetry]
name = "cannot_get_latest_package"
version = "0.1.0"
description = ""
authors = ["Tamás Szelei <szelei.t@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
conan = "1.21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
